parse-events: Handle invalid opcode parsing gracefully
Patchwork-id: 49801
O-Subject: [PATCH RHEL6.4 BZ846702 005/198] parse-events: Handle invalid opcode parsing gracefully
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit d69afed55be1016c2bcfcb3e00cd5365d2f557f6
Author: Vaibhav Nagarnaik <vnagarnaik@google.com>
Date:   Fri Apr 6 00:48:00 2012 +0200

description
===========
If an invalid opcode is encountered, trace-cmd exits with an error.
Instead it can be treated as a soft error where the event's print format
is not parsed and its binary data is dumped out.

This patch adds a return value to arg_num_eval() function to indicate if
the parsing was successful. If not, then the error is considered soft
and the parsing of the offending event fails.

Signed-off-by: Jarod Wilson <jarod@redhat.com>