perf: Have perf use the new libtraceevent.a library
Patchwork-id: 49813
O-Subject: [PATCH RHEL6.4 BZ846702 016/198] perf: Have perf use the new libtraceevent.a library
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit aaf045f72335653b24784d6042be8e4aee114403
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Fri Apr 6 00:47:56 2012 +0200

description
===========
The event parsing code in perf was originally copied from trace-cmd
but never was kept up-to-date with the changes that was done there.
The trace-cmd libtraceevent.a code is much more mature than what is
currently in perf.

This updates the code to use wrappers to handle the calls to the
new event parsing code. The new code requires a handle to be pass
around, which removes the global event variables and allows
more than one event structure to be read from different files
(and different machines).

But perf still has the old global events and the code throughout
perf does not yet have a nice way to pass around a handle.
A global 'pevent' has been made for perf and the old calls have
been created as wrappers to the new event parsing code that uses
the global pevent.

With this change, perf can later incorporate the pevent handle into
the perf structures and allow more than one file to be read and
compared, that contains different events.