traceevent: Let filtering numbers by string use function names
Patchwork-id: 49829
O-Subject: [PATCH RHEL6.4 BZ846702 035/198] tools lib traceevent: Let filtering numbers by string use function names
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit e84c282b40251f314c429f39b044785e323f2648
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Tue May 22 14:45:21 2012 +0900

description
===========
As a pointer can be converted into a function name, let the filters
work with the function name as well as with the pointer number.  If
the comparison expects a string, then convert numbers into functions,
but only when the number is the same size as a long.