traceevent: Add support for "%.*s" in bprintk events
Patchwork-id: 49834
O-Subject: [PATCH RHEL6.4 BZ846702 036/198] tools lib traceevent: Add support for "%.*s" in bprintk events
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit c2e6dc2b268cca44d522b2ee86147f0d30d7e3e4
Author: Steven Rostedt <srostedt@redhat.com>
Date:   Tue Nov 15 18:47:48 2011 -0500

description
===========
The arg notation of '*' in bprintks is not handled by the parser.
Implement it so that they show up properly in the output and do not
kill the tracer from reporting events.