perf ui browser: Return the current color when setting a new one
Patchwork-id: 49844
O-Subject: [PATCH RHEL6.4 BZ846702 053/198] perf ui browser: Return the current color when setting a new one
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit 087091652bf8b351432a3f3fb50996ee3582d5e4
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Mon Apr 2 12:48:56 2012 -0300

description
===========
Tools that want to change parts of the line to a different color and
then restore the previous one will use this, starting with the annotate
browser that will change the color of addresses if not on the current
entry, i.e. the selected one.