perf report: Correct display of samples and events in header
Patchwork-id: 49868
O-Subject: [PATCH RHEL6.4 BZ846702 059/198] perf report: Correct display of samples and events in header
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit cc68628096b813010f30f69fcf7d1832d18fa767
Author: Ashay Rane <ashay.rane@tacc.utexas.edu>
Date:   Thu Apr 5 21:01:01 2012 -0500

description
===========
This patch prints the number of samples and the count of performance
events separately.

This allows comparing performance of different applications with each
other.

Previously, the sample count was displayed against an 'Events:' heading.
With this patch, the header now reads (for example):

Samples: 5K of event 'instructions'
Event count (approx.): 2993026545

The patch covers both the stdio and the browser interface.