perf annotate: Mark jump instructions with no offset
Patchwork-id: 49867
O-Subject: [PATCH RHEL6.4 BZ846702 091/198] perf annotate: Mark jump instructions with no offset
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit fb29fa58e36df09c807d252247d64a221fcd5bbb
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed Apr 25 14:16:03 2012 -0300

description
===========
I.e. jumps that go to code outside the current function, that is denoted
in objdump -dS as:

   399f877a9f: jne    399f87bcf4 <_L_lock_5154>

I.e. without the + after the name of the current function, like in:

   399f877aa5: jmp    399f877ab2 <_int_free+0x412>

The browser will use that info to avoid drawing connectors to the start
of the function, since ops.target.addr was zero.