perf evlist: Explicititely initialize input_name
Patchwork-id: 49917
O-Subject: [PATCH RHEL6.4 BZ846702 161/198] perf evlist: Explicititely initialize input_name
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit a1d44b9acdf7b2820ef762846150b4d13b84e075
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed May 23 21:47:51 2012 -0300

description
===========
It was a global variable, so it was initialized, implicitely, to zero by
being placed in the bss.

Now it is just a local variable that is then passed to the __cmd_evlist
routine, so it must be explicitely set to NULL.

The problem manifested on a Fedora 17 system, using:

 gcc version 4.7.0 20120507 (Red Hat 4.7.0-5) (GCC)

But not on several other systems, by luck.