perf report: Use the right symbol for annotation
Patchwork-id: 49939
O-Subject: [PATCH RHEL6.4 BZ846702 177/198] perf report: Use the right symbol for annotation
Bugzilla: 846702
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 846702
https://bugzilla.redhat.com/show_bug.cgi?id=846702

upstream status
===============
commit 91557e847dbc715acf2d847a1ffec63f71b00b65
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Wed May 30 12:25:52 2012 -0300

description
===========
In non symbolic views, i.e. --sort without "symbol", as in:

 perf report --sort comm

We're segfaulting in the --tui because we're testing the symbol resolved
and then trying to use the symbol on the histogram entry where we're
coalescing all hits for a COMM, and the first hist_entry for a comm may
have a NULL symbol, i.e. the RIP didn't resolve to any symbol.

In this case we're segfaulting, fix it by testing against the symbol in
the histogram entry.