mptfusion: Fix for device removed in blocked state
Patchwork-id: 50386
O-Subject: [RHEL6.4 PATCH 1/1] mptfusion: Fix for issue - The device is removed in blocked state
Bugzilla: 831865
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

This patch comes from LSI, a customer tested this fix
and it solved his issue with a device which entered blocked state.
scsi mid level will not dispatch the command(scsi_dispatch_cmd)
since the device is in blocked state. A timeout 'task.. blocked for
more than 120 seconds.' can then be seen.

brew build - task_4795884/

Tomas

Upstream post:
http://marc.info/?l=linux-scsi&m=134269193618776&w=2

Fix for issue - The device is removed in blocked state.

The device is set to blocked state when I/Os running and the cable is pulled.
Also if the device missing delay timer is started. When the device missing
delay timer expires, the device is removed but the device state is not set to
unblocked state. This will cause problems  if there are commands in block
queue.

To fix this issue, added code to unblock the device before removing it.

Signed-off-by: Sreekanth Reddy <sreekanth.reddy@lsi.com>
Cc: <stable@kernel.org>
Signed-off-by: Jarod Wilson <jarod@redhat.com>