perf: Add support to parse event group syntax
Patchwork-id: 50482
O-Subject: [PATCH RHEL6.4 BZ695765 5/8] perf tools: Add support to parse event group syntax
Bugzilla: 695765
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 695765
https://bugzilla.redhat.com/show_bug.cgi?id=695765

Upstream status:
================
commit 89efb029502d7f2d0993ed2aa9280c414336939c
Author: Jiri Olsa <jolsa@redhat.com>
Date:   Wed Aug 8 12:14:14 2012 +0200

Description:
============
Adding scanner/parser bits to parse event groups.

The grammar for group is:
  groups:      groups ',' group | group
  group:       group_name '{' events '}' group_mod
  group_name:  name | empty
  group_mod:   ':' group_mods | empty
  group_mods:  event_mod

It's possible to use standard event modifier as a modifier
for group. It'll be used as an update to existing event
modifiers.

It's necessary to use quoting ("'\) when specifying group on
command line, since {} characters are interpreted by most of
the shells.

It is now possible to specify groups in event syntax like:

  '{cycles,faults}'
   - anonymous group

  'group1{cycles,faults}
   - group with name 'group1'

  '{cycles,faults}:k
   - anonymous group with event modifier 'k'

  '{cpu-clock,task-clock},{minor-faults,major-faults}'
   - two anonymous groups

The grouping functionality itself is coming shortly.

Signed-off-by: Jarod Wilson <jarod@redhat.com>