perf: Add support to update event modifier
Patchwork-id: 50486
O-Subject: [PATCH RHEL6.4 BZ695765 6/8] perf tools: Add support to update event modifier
Bugzilla: 695765
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 695765
https://bugzilla.redhat.com/show_bug.cgi?id=695765

Upstream status:
================
commit f5b1135bf79557563a814e53ecd610cce663c1e3
Author: Jiri Olsa <jolsa@redhat.com>
Date:   Wed Aug 8 12:21:54 2012 +0200

Description:
============
Adding support to update already defined event's attribute with
event modifier. This change will allow to use group modifier as
an update to the existing event modifiers.

Adding 'add' parameter to the parse_events__modifier_event function.
Calling it with 'add' = false/true, the event modifier is
initialized/updated respectively.

Added exclude_GH flag to evsel struct, because we need to remember
if one of 'GH' modifiers was used for event. The reason is that the
default settings for exclude_guest is 1 and during the group
modifier processing we have no other way of knowing if it was set
by default or by event modifier.

Keeping the current behaviour, that any event/group modifier reset
the defaults for exclude_host (0) and exclude_guest (1).

Signed-off-by: Jarod Wilson <jarod@redhat.com>