core: fix deadlock in bConfigurationValue attribute method
Patchwork-id: 50511
O-Subject: [RHEL6 PATCH 2/2] USB: fix deadlock in bConfigurationValue attribute method
Bugzilla: 832126
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=832126

commit 8963c487a80b4688c9e68dcc504a90074aacc145
Author: Alan Stern <stern@rowland.harvard.edu>
Date:   Tue Apr 17 15:22:39 2012 -0400

    USB: fix deadlock in bConfigurationValue attribute method

    This patch (as154) fixes a self-deadlock that occurs when userspace
    writes to the bConfigurationValue sysfs attribute for a hub with
    children.  The task tries to lock the bandwidth_mutex at a time when
    it already owns the lock:

    	The attribute's method calls usb_set_configuration(),
    	which calls usb_disable_device() with the bandwidth_mutex
    	held.

    	usb_disable_device() unregisters the existing interfaces,
    	which causes the hub driver to be unbound.

    	The hub_disconnect() routine calls hub_quiesce(), which
    	calls usb_disconnect() for each of the hub's children.

    	usb_disconnect() attempts to acquire the bandwidth_mutex
    	around a call to usb_disable_device().

    The solution is to make usb_disable_device() acquire the mutex for
    itself instead of requiring the caller to hold it.  Then the mutex can
    cover only the bandwidth deallocation operation and not the region
    where the interfaces are unregistered.

    This has the potential to change system behavior slightly when a
    config change races with another config or altsetting change.  Some of
    the bandwidth released from the old config might get claimed by the
    other config or altsetting, make it impossible to restore the old
    config in case of a failure.  But since we don't try to recover from
    config-change failures anyway, this doesn't matter.

    [This should be marked for stable kernels that contain the commit
    fccf4e86200b8f5edd9a65da26f150e32ba79808 "USB: Free bandwidth when
    usb_disable_device is called."
    That commit was marked for stable kernels as old as 2.6.32.]

    Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
    Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
    Cc: stable <stable@vger.kernel.org>
    Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

v2:  Thanks to Nikolay's sharp eyes, he noticed a possible deadlock.  This is the
results of shifting upstream code.  I figured out the patches needed to sync all this
up, but it is going to be confusing to explain.  So here goes:

(upstream commits)
0791971 usb: allow drivers to use allocated bandwidth until unbound

[posted in previous patch of this series] fixes an issue with using the
driver's bandwidth until unbound.  Most importantly was moving the
disable_device outside the bandwidth_mutex.  This caused a few issues,
which were cleaned up with:

ca5c485 USB: additional regression fix for device removal
e534c5b USB: fix regression occurring during device removal
fccf4e8 USB: Free bandwidth when usb_disable_device is called.

However those fixes were included in RHEL-6.2.  But fccf4e8 was modified
when backported.  It didn't bring in a mutex change because it wasn't need
as 0791971 was not backported at the time.

Now that I am backporting 0791971 [previous patch], the mutex change comes into
play, but _this_ patch moves it back to the original spot (because a proper locking
clean-up is in place).

So when reviewing this patch and you notice the bandwith_mutex movement is missing,
it is because it was a no-op from a previous dropped backport.

Whew.

The deadlock Nikolay noticed should be gone now and the code more aligned with upstream
for now.

[maintaining a stack is much more a PITA than a driver :-) ]

Feel to ask any questions if my above explaination is confusing.

Signed-off-by: Jarod Wilson <jarod@redhat.com>