ext3, ext4: don't clear orphan list on ro mount with errors
Patchwork-id: 50518
O-Subject: [PATCH RHEL6.4] ext3/ext4: don't clear orphan list on ro mount with errors
Bugzilla: 852771
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

This is for:

Bug 850803 - Ext3/Ext4 orphan inode list cleared if mounted after IO error

The following patches for ext3 & ext4 were submitted upstream & acked
by the maintainers, though they are not yet merged into Linus' tree.

===

When we have a filesystem with an orphan inode list *and* in error
state, things behave differently if:

1) e2fsck -p is done prior to mount: e2fsck fixes things and exits
   happily (barring other significant problems)

vs.

2) mount is done first, then e2fsck -p: due to the orphan inode
   list removal, more errors are found and e2fsck exits with
   UNEXPECTED INCONSISTENCY.

The 2nd case above, on the root filesystem, has the tendency to halt
the boot process, which is unfortunate.

The situation can be improved by not clearing the orphan
inode list when the fs is mounted readonly.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>