dm-raid: Move some code to its own function
Patchwork-id: 50565
O-Subject: [RHEL6.4 PATCH 11 OF 13] DM RAID: Move some code to its own function
Bugzilla: 854058
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bug 854058

Staged for upstream:
http://git.neil.brown.name/git?p=md.git;a=commit;h=4ec91f095986bcfefcff81b55312c2c789ab7fcc

DM RAID:  Move chunk of code to it's own function

The code that checks whether device replacements/rebuilds are possible given
a specific RAID type is moved to it's own function.  It will further expand
when the code to check RAID10 is added.  A separate function makes it easier
to read.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>