hpsa: fix physical device lun and target numbering problem
Patchwork-id: 50584
O-Subject: [RHEL6.4 PATCH 06/37] hpsa: fix physical device lun and target numbering problem
Bugzilla: 737644
RH-Acked-by: Rob Evers <revers@redhat.com>

[SCSI] hpsa: fix physical device lun and target numbering problem

Upstream: 01350d05539d1c95ef3568d062d864ab76ae7670

If a physical device exposed to the OS by hpsa
is replaced (e.g. one hot plug tape drive is replaced
by another, or a tape drive is placed into "OBDR" mode
in which it acts like a CD-ROM device) and a rescan is
initiated, the replaced device will be added to the
SCSI midlayer with target and lun numbers set to -1.
After that, a panic is likely to ensue.  When a physical
device is replaced, the lun and target number should be
preserved.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>
Cc: stable@kernel.org
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>