hpsa: do not skip disabled devices
Patchwork-id: 50595
O-Subject: [RHEL6.4 PATCH 20/37] hpsa: do not skip disabled devices
Bugzilla: 737644
RH-Acked-by: Rob Evers <revers@redhat.com>

[SCSI] hpsa: do not skip disabled devices

Upstream: e754b42424e0bd5949f47118f71720c42b93c6e0

There was code to skip "disabled" devices which was intended to
skip devices disabled in the BIOS, but it really just checks to
see if the device can write to host memory, which this is disabled
by pci_disable_device on driver unload, so this check has the effect
of preventing subsequent load of the driver.  And devices disabled in
the BIOS don't show up at all anyway, so this check never made any
sense to begin with, and should be removed.

Signed-off-by: Stephen M. Cameron <scameron@beardog.cce.hp.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>