hpsa: do aborts two ways
Patchwork-id: 50601
O-Subject: [RHEL6.4 PATCH 28/37] hpsa: do aborts two ways
Bugzilla: 737644
RH-Acked-by: Rob Evers <revers@redhat.com>

[SCSI] hpsa: do aborts two ways

Upstream: 6cba3f1972de14421ef4cf4b46a15cc5d604e791

When aborting a command, the tag is supposed to be
specified as 64-bit little endian.  However, some smart
arrays expect the tag of the command to be aborted to be
specified in a strange byte order.  How to tell which sort
of Smart Array firmware we're dealing with is not obvious.
However, because of the way we construct our tags, the values
of any outstanding tag when specified with the "strange" byte
order will not collide with the value specified in the correct
order.  That means we can safely attempt the abort both ways.

Signed-off-by: Stephen M. Cameron <stephenmcameron@gmail.com>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>