nfs: Don't use DATA_SYNC writes
Patchwork-id: 50655
O-Subject: [RHEL6.4 PATCH 14/18] NFS: Don't use DATA_SYNC writes
Bugzilla: 854746
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

If we're writing back data, and the FLUSH_STABLE flag is set, then we
always want to use NFS_FILE_SYNC, since we're always in a situation where
we're doing page reclaim, and so we want to free up the page as quickly
as possible.

If we're in the FLUSH_COND_STABLE case, then we either want to use another
unstable write (if we have to do a commit anyway) or again, we want to
use NFS_FILE_SYNC because we know that we have no more pages to write
out.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream Id: 87ed5eb44ad9338b1617a0e78dea81d681325298

Signed-off-by: Jarod Wilson <jarod@redhat.com>