xfs: push the AIL from memory reclaim and periodic sync
Patchwork-id: 50892
O-Subject: [RHEL6.4 PATCH] xfs: push the AIL from memory reclaim and periodic sync
Bugzilla: 855139
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Brian Foster <bfoster@redhat.com>

Upstream commit: fd074841cfe01b006465fb9388091012585e8dfb
RH BZ: 855139

When we are short on memory, we want to expedite the cleaning of
dirty objects.  Hence when we run short on memory, we need to kick
the AIL flushing into action to clean as many dirty objects as
quickly as possible.  To implement this, sample the lsn of the log
item at the head of the AIL and use that as the push target for the
AIL flush.

Further, we keep items in the AIL that are dirty that are not
tracked any other way, so we can get objects sitting in the AIL that
don't get written back until the AIL is pushed. Hence to get the
filesystem to the idle state, we might need to push the AIL to flush
out any remaining dirty objects sitting in the AIL. This requires
the same push mechanism as the reclaim push.

This patch also renames xfs_trans_ail_tail() to xfs_ail_min_lsn() to
match the new xfs_ail_max_lsn() function introduced in this patch.
Similarly for xfs_trans_ail_push -> xfs_ail_push.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Reviewed-by: Alex Elder <aelder@sgi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>