mpt2sas: Set the phy identifier of the end device to the phy number of the parent device it is linked to
Patchwork-id: 50910
O-Subject: [RHEL6.4 PATCH 13/16] mpt2sas: Set the phy identifier of the end device to the phy number of the parent device it is linked to
Bugzilla: 826414
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Upstream: 64bb813 [SCSI] mpt2sas: Set the phy identifier of the end device to the phy number of the parent device it is linked to

The phy_identifier inside the routine _transport_set_identify()
is set to sas_device_page_zero->PhyNum. This returns the
phy number of the parent device this device is linked to.

Signed-off-by: Nagalakshmi Nandigama <nagalakshmi.nandigama@lsi.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>