fix contig_page_data kABI breakage and related memory corruption
Patchwork-id: 51070
O-Subject: [RHEL6.4 PATCH 1/6] mm: fix contig_page_data kABI breakage
Bugzilla: 853007
RH-Acked-by: Frantisek Hrbata <fhrbata@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla
========
BZ#853007
https://bugzilla.redhat.com/show_bug.cgi?id=853007

Test Status
===========
Tested on Hitachi HA8000-tc server.
I confirmed that the issue written in BZ853007 doesn't happen anymore.

Description
===========
It was discovered that we silently broke kabi in 6.3.

This commit modified an enum struct in include/linux/mmzone.h

commit cb08c8ccb678e8b9fa90a630a92f9eeac6cc70d8
Author: Satoru Moriya <smoriya@redhat.com>
Date:   Wed Feb 15 00:43:55 2012 -0500

    [mm] vmscan: do not writeback filesystem pages in direct reclaim

    #--- a/include/linux/mmzone.h
    #+++ b/include/linux/mmzone.h
    @@ -100,6 +100,7 @@ enum zone_stat_item {
            NR_UNSTABLE_NFS,        /* NFS unstable pages */
            NR_BOUNCE,
            NR_VMSCAN_WRITE,
    +       NR_VMSCAN_WRITE_SKIP,
            NR_WRITEBACK_TEMP,	/* Writeback using temporary buffers */
            NR_ISOLATED_ANON,	/* Temporary isolated pages from anon lru */
            NR_ISOLATED_FILE,	/* Temporary isolated pages from file lru */

Now this should have broken kabi for the symbol 'contig_page_data' as the
'struct zone' allocation (which is based on the number of elements in the
enum) would have grown by one causing the breakage.  However the kabi
checker doesn't fully understand enum sizes and did not pick this up.

[That is fixed upstream (e37ddb825) but we can't backport it because it
will rewrite all of our checksums...]

As a result a partner stumbled upon this breakage and requested we fix it
as it changes the struct element offsets for 32-bit (not 64-bit).

The fix is simply to revert the above piece.

This patch removes the code which increases vm_stat length. With this
patch, we can't find how many pages skip writeback in page reclaim
path from procfs, that also added the patchset above.

Signed-off-by: Satoru Moriya <smoriya@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>