gfs2: Change variable blk to biblk
Patchwork-id: 51121
O-Subject: [PATCH 07/21] GFS2: Change variable blk to biblk
Bugzilla: 753337
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

In the resource group code, we have no less than three different
kinds of block references: block relative to the file system (u64),
block relative to the rgrp (u32), and block relative to the bitmap.
This is a small step to making the code more readable; it renames
variable blk to biblk to solidify in my mind that it's relative to
the bitmap and nothing else.

rhbz#753337

Signed-off-by: Jarod Wilson <jarod@redhat.com>