gfs2: Use variable rather than qa to determine if unstuff necessary
Patchwork-id: 51124
O-Subject: [PATCH 10/21] GFS2: Use variable rather than qa to determine if unstuff necessary
Bugzilla: 753337
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

In the future, the qadata structure will be eliminated and merged
back in with the block reservation structure, after we extend the
lifespan of that. This patch is a step forward in eliminating the
qadata structure. It adds a variable to the do_grow function to
determine when unstuffing is necessary, and has been done.

rhbz#753337

Signed-off-by: Jarod Wilson <jarod@redhat.com>