gfs2: Extend the life of the reservations structure
Patchwork-id: 51128
O-Subject: [PATCH 15/21] GFS2: Extend the life of the reservations structure
Bugzilla: 753337
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

This patch lengthens the lifespan of the reservations structure for
inodes. Before, they were allocated and deallocated for every write
operation. With this patch, they are allocated when the first write
occurs, and deallocated when the last process closes the file.
It's more efficient to do it this way because it saves GFS2 a lot of
unnecessary allocates and frees. It also gives us more flexibility
for the future: (1) we can now fold the qadata structure back into
the structure and save those alloc/frees, (2) we can use this for
multi-block reservations.

rhbz#753337

Signed-off-by: Jarod Wilson <jarod@redhat.com>